+++ src/classes/features/achievementTracker/swapAchievement.ts
@@ -1513,6 +1513,16 @@ function getSwappedUnlockPillEffect(unlock: Unlock): Unlock | undefined {
 function getSwappedUnlockHeart(unlock: Unlock): Unlock | undefined {
   const heartUnlock = unlock as HeartUnlock;
 
+  // 18
+  if (!isRoomTypeUnlocked(RoomType.CLEAN_BEDROOM, false)) {
+    return getUnlock(UnlockType.ROOM, RoomType.CLEAN_BEDROOM);
+  }
+
+  // 19
+  if (!isRoomTypeUnlocked(RoomType.DIRTY_BEDROOM, false)) {
+    return getUnlock(UnlockType.ROOM, RoomType.CLEAN_BEDROOM);
+  }
+
   if (isHardcoreMode()) {
     const worseHeartSubType = getWorseLockedHeartSubType(
       heartUnlock.heartSubType,
@@ -1621,6 +1631,23 @@ const SWAPPED_UNLOCK_OTHER_FUNCTIONS = new ReadonlyMap<
   OtherUnlockKind,
   () => Unlock | undefined
 >([
+  [
+    OtherUnlockKind.BEDS,
+    () => {
+      // 18
+      if (!isRoomTypeUnlocked(RoomType.CLEAN_BEDROOM, false)) {
+        return getUnlock(UnlockType.ROOM, RoomType.CLEAN_BEDROOM);
+      }
+
+      // 19
+      if (!isRoomTypeUnlocked(RoomType.DIRTY_BEDROOM, false)) {
+        return getUnlock(UnlockType.ROOM, RoomType.CLEAN_BEDROOM);
+      }
+
+      return undefined;
+    },
+  ],
+
   [
     OtherUnlockKind.BLUE_FIREPLACES,
     () => {