+++ src/enums/BossIDCustom.ts
@@ -11,6 +11,10 @@ export const BossIDCustom = {
   KRAMPUS: (HIGHEST_BOSS_ID + 1) as BossID, // 81.1 / MinibossID.KRAMPUS (15)
   URIEL: (HIGHEST_BOSS_ID + 2) as BossID, // 271
   GABRIEL: (HIGHEST_BOSS_ID + 3) as BossID, // 272
+  ULTRA_FAMINE: (HIGHEST_BOSS_ID + 4) as BossID, // 951.10
+  ULTRA_PESTILENCE: (HIGHEST_BOSS_ID + 5) as BossID, // 951.20
+  ULTRA_WAR: (HIGHEST_BOSS_ID + 6) as BossID, // 951.30
+  ULTRA_DEATH: (HIGHEST_BOSS_ID + 7) as BossID, // 951.40
 } as const;
 
 validateCustomEnum("BossIDCustom", BossIDCustom);
@@ -27,6 +31,22 @@ export function getBossNameCustom(bossID: BossID): string {
 
     case BossIDCustom.GABRIEL: {
       return "Gabriel";
+    }
+
+    case BossIDCustom.ULTRA_FAMINE: {
+      return "Ultra Famine";
+    }
+
+    case BossIDCustom.ULTRA_PESTILENCE: {
+      return "Ultra Pestilence";
+    }
+
+    case BossIDCustom.ULTRA_WAR: {
+      return "Ultra War";
+    }
+
+    case BossIDCustom.ULTRA_DEATH: {
+      return "Ultra Death";
     }
 
     default: {